- Reduce size further - In launcher debug mode, report received buffer size (strangely this is high, will investigate further)
SVN r4150 (trunk)

+++ odalpapi/net_packet.h
@@ -91,7 +91,7 @@ struct Cvar_t
 	std::string Name;
     std::string Value;
 
-	int32_t Type;
+	int8_t Type;
 
 	union
 	{
